requirements.txt
fire
tqdm
openpyxl
pandas
scikit-learn
numpy<2.0
matplotlib
seaborn
pydantic==1.9
tensorflow==2.19.1
tensorflow-text==2.19.0
tensorflow-hub
wandb==0.21.3
protobuf<=3.20.3
openai
tiktoken
python-dotenv
